requirements.txt
fastapi~=0.95.1
loguru~=0.7.0
SQLAlchemy~=2.0.13
pydantic~=1.10.7
aioredis~=2.0.1
uvicorn~=0.22.0
passlib~=1.7.4
python-jose~=3.3.0
